refactor(datasets): break contents of schema into query and mutation sub-modules (#3239)
