+++ app/schema.graphql
@@ -565,6 +565,11 @@ type Model {
 }
 
 type Mutation {
+  deleteProject(id: GlobalID!): Query!
+  clearProject(id: GlobalID!): Query!
+  createDataset(name: String!, description: String = null, metadata: JSON = null): Dataset!
+  addSpansToDataset(input: AddSpansToDatasetInput!): AddSpansToDatasetPayload!
+
   """
   Given a list of event ids, export the corresponding data subset in Parquet format. File name is optional, but if specified, should be without file extension. By default the exported file name is current timestamp.
   """
@@ -574,10 +579,6 @@ type Mutation {
   Given a list of clusters, export the corresponding data subset in Parquet format. File name is optional, but if specified, should be without file extension. By default the exported file name is current timestamp.
   """
   exportClusters(clusters: [ClusterInput!]!, fileName: String): ExportedFile!
-  deleteProject(id: GlobalID!): Query!
-  clearProject(id: GlobalID!): Query!
-  createDataset(name: String!, description: String = null, metadata: JSON = null): Dataset!
-  addSpansToDataset(input: AddSpansToDatasetInput!): AddSpansToDatasetPayload!
 }
 
 """An object with a Globally Unique ID"""
