+++ src/phoenix/server/api/mutations/__init__.py
@@ -1,0 +1,10 @@
+import strawberry
+
+from phoenix.server.api.mutations.dataset_mutations import DatasetMutationMixin
+from phoenix.server.api.mutations.export_events_mutations import ExportEventsMutationMixin
+from phoenix.server.api.mutations.project_mutations import ProjectMutationMixin
+
+
+@strawberry.type
+class Mutation(ProjectMutationMixin, DatasetMutationMixin, ExportEventsMutationMixin):
+    pass
